Script that finds the nearest pickable object, added a "Pickable" tag to ores
What the title says, plus, there is still left to actually check if an item is being held and a method that checks if the ore is the one that is being picked. I will then add a toggle feature to allow picking up and dropping of items with a button click (Letter E).